+++ BetterContinents.cs
@@ -49,6 +49,8 @@ public partial class BetterContinents : BaseUnityPlugin
 
     public static ConfigEntry<string> ConfigSpawnFile;
 
+    public static ConfigEntry<string> ConfigTerrainFile;
+    public static ConfigEntry<string> ConfigDefaultTerrainColor;
     public static ConfigEntry<string> ConfigPaintFile;
     public static ConfigEntry<string> ConfigLavaFile;
     public static ConfigEntry<string> ConfigMossFile;
@@ -189,6 +191,15 @@ public partial class BetterContinents : BaseUnityPlugin
                 groupBuilder.AddValue("Biome precision")
                     .Description("Adjusts how precisely terrain is matched to the biomemap (0 = vanilla, 1 = 3x3, 2 = 5x5, etc.)")
                     .Default(0).Range(0, 5).Bind(out ConfigBiomePrecision);
+            })
+            .AddGroup("BetterContinents.Terrainmap", groupBuilder =>
+            {
+                groupBuilder.AddValue("Terrainmap file")
+                    .Description("Path to a terrainmap file to use. See thea description on Nexusmods.com for the specifications (it will fail if they are not met)")
+                    .Default("").Bind(out ConfigTerrainFile);
+                groupBuilder.AddValue("Default Terrain Color")
+                    .Description("Terrainmap color that uses the default terrain generation")
+                    .Default("#808080").Bind(out ConfigDefaultTerrainColor);
             })
             .AddGroup("BetterContinents.Forest", groupBuilder =>
             {