+++ ImageMapColor.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections.Generic;
-using System.Diagnostics;
+﻿using System.Diagnostics;
 using SixLabors.ImageSharp;
 using SixLabors.ImageSharp.PixelFormats;
 using UnityEngine;
@@ -8,13 +7,14 @@ namespace BetterContinents;
 
 internal class ImageMapColor() : ImageMapBase()
 {
-    public static ImageMapColor? Create(string path)
+    public static ImageMapColor? Create(string path, Rgba32? defaultColor)
     {
         if (string.IsNullOrEmpty(path))
             return null;
         ImageMapColor map = new()
         {
-            FilePath = path
+            FilePath = path,
+            DefaultColor = defaultColor
         };
         if (!map.LoadSourceImage())
             return null;
@@ -22,19 +22,22 @@ internal class ImageMapColor() : ImageMapBase()
             return null;
         return map;
     }
-    public static ImageMapColor? Create(byte[] data, string path)
+    public static ImageMapColor? Create(byte[] data, string path, Rgba32? defaultColor)
     {
         ImageMapColor map = new()
         {
             FilePath = path,
-            SourceData = data
+            SourceData = data,
+            DefaultColor = defaultColor
         };
         if (!map.CreateMap())
             return null;
         return map;
     }
-    public static ImageMapColor? Create(byte[] data) => Create(data, "");
+    public static ImageMapColor? Create(byte[] data, Rgba32? defaultColor) => Create(data, "", defaultColor);
+    public Rgba32? DefaultColor;
     private UnityEngine.Color[] Map = [];
+    private UnityEngine.Color?[] NullableMap = [];
 
     public bool CreateMap() => CreateMap<Rgba32>();
     protected override bool LoadTextureToMap<T>(Image<T> image)
@@ -42,11 +45,24 @@ internal class ImageMapColor() : ImageMapBase()
         var st = new Stopwatch();
         st.Start();
 
-        var colorMapping = new Dictionary<Color32, Heightmap.Biome>(new Color32Comparer());
         var img = (Image<Rgba32>)(Image)image;
-        Map = LoadPixels(img, pixel => new UnityEngine.Color(pixel.R / 255f, pixel.G / 255f, pixel.B / 255f, pixel.A / 255f));
+        if (DefaultColor == null)
+        {
+            Map = LoadPixels(img, pixel => new UnityEngine.Color(pixel.R / 255f, pixel.G / 255f, pixel.B / 255f, pixel.A / 255f));
+        }
+        else
+        {
+            NullableMap = LoadPixels(img, pixel => (UnityEngine.Color?)new UnityEngine.Color(pixel.R / 255f, pixel.G / 255f, pixel.B / 255f, pixel.A / 255f));
+            var defaultColor = new UnityEngine.Color(DefaultColor.Value.R / 255f, DefaultColor.Value.G / 255f, DefaultColor.Value.B / 255f, DefaultColor.Value.A / 255f);
+            for (int i = 0; i < NullableMap.Length; i++)
+            {
+                if (NullableMap[i] == defaultColor)
+                    NullableMap[i] = null;
+            }
+        }
 
-        BetterContinents.Log($"Time to calculate paints from {FilePath}: {st.ElapsedMilliseconds} ms");
+
+        BetterContinents.Log($"Time to calculate colors from {FilePath}: {st.ElapsedMilliseconds} ms");
         return true;
     }
 
@@ -56,7 +72,7 @@ internal class ImageMapColor() : ImageMapBase()
         var path = pkg.ReadString();
         if (string.IsNullOrEmpty(path))
             return null;
-        return Create(pkg.ReadByteArray(), path);
+        return Create(pkg.ReadByteArray(), path, null);
     }
 
 
@@ -80,9 +96,36 @@ internal class ImageMapColor() : ImageMapBase()
         var p10 = Map[y0 * Size + x1];
         var p01 = Map[y1 * Size + x0];
         var p11 = Map[y1 * Size + x1];
-
         var a = UnityEngine.Color.Lerp(p00, p10, xd);
         var b = UnityEngine.Color.Lerp(p01, p11, xd);
         return UnityEngine.Color.Lerp(a, b, yd);
+    }
+    public bool TryGetValue(float x, float y, ref UnityEngine.Color color)
+    {
+        float xa = x * (Size - 1);
+        float ya = y * (Size - 1);
+
+        int xi = Mathf.FloorToInt(xa);
+        int yi = Mathf.FloorToInt(ya);
+
+        float xd = xa - xi;
+        float yd = ya - yi;
+
+        int x0 = Mathf.Clamp(xi, 0, Size - 1);
+        int x1 = Mathf.Clamp(xi + 1, 0, Size - 1);
+        int y0 = Mathf.Clamp(yi, 0, Size - 1);
+        int y1 = Mathf.Clamp(yi + 1, 0, Size - 1);
+
+        var p00 = NullableMap[y0 * Size + x0];
+        var p10 = NullableMap[y0 * Size + x1];
+        var p01 = NullableMap[y1 * Size + x0];
+        var p11 = NullableMap[y1 * Size + x1];
+        if (p00 == null || p10 == null || p01 == null || p11 == null)
+            return false;
+
+        var a = UnityEngine.Color.Lerp((UnityEngine.Color)p00, (UnityEngine.Color)p10, xd);
+        var b = UnityEngine.Color.Lerp((UnityEngine.Color)p01, (UnityEngine.Color)p11, xd);
+        color = UnityEngine.Color.Lerp(a, b, yd);
+        return true;
     }
 }